+++ lib/sqflite_live.dart
@@ -5,3 +5,4 @@ library;
 
 export './src/database_extension.dart';
 export './src/host/logger/log_me.dart' show LogLevel;
+export './src/host/sqflite_live.dart' show SqfliteLive;